refactor(ui): split app.py below the 800-line cap
The health API is the one part of app.py that answers to a probe rather
than to a browser: three unauthenticated routes, a dependency probe, the
redaction rules that keep person names and host paths out of a public
payload, and the snapshot cache that stops a probe flood from becoming a
work flood. None of it touches the wizard pages, the sidebar, or the
session machinery around them, and nothing left in app.py calls into it.

It moves to ui/health_api.py verbatim and registers through
register_health_routes(app), mirroring register_trigger_routes(app) on
the next line. app.py keeps the registration call, so route order is
unchanged - verified by dumping app.routes and app.user_middleware
before and after: identical but for the middleware object's address.

app.py 891 -> 629; health_api.py 290.

Part of #686
