+++ tests/test_health.py
@@ -40,11 +40,11 @@ def _fresh_health_snapshot():
 
     Tests would otherwise read each other's snapshots, so each starts cold.
     """
-    from immich_memories.ui import app as app_module
+    from immich_memories.ui import health_api as health_module
 
-    app_module._health_snapshot_cache = None
+    health_module._health_snapshot_cache = None
     yield
-    app_module._health_snapshot_cache = None
+    health_module._health_snapshot_cache = None
 
 
 class TestHealthEndpoint:
@@ -60,7 +60,7 @@ class TestHealthEndpoint:
         expected_diagnostic: str,
     ) -> None:
         """Optional status failures retain class diagnostics without their sensitive message."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = _config_with_every_secret_class()
         sensitive_values = (config.immich.url, *configured_secret_values(config))
@@ -73,10 +73,10 @@ class TestHealthEndpoint:
             last_run.side_effect = failure
 
         with (
-            caplog.at_level(logging.WARNING, logger="immich_memories.ui.app"),
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            caplog.at_level(logging.WARNING, logger="immich_memories.ui.health_api"),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(
                     status="ready",
@@ -84,8 +84,8 @@ class TestHealthEndpoint:
                     resolved_api_version="v3",
                 ),
             ),
-            patch("immich_memories.ui.app._get_automation_status", automation),
-            patch("immich_memories.ui.app._get_last_successful_run", last_run),
+            patch("immich_memories.ui.health_api._get_automation_status", automation),
+            patch("immich_memories.ui.health_api._get_last_successful_run", last_run),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -101,7 +101,7 @@ class TestHealthEndpoint:
 
         client = TestClient(app, raise_server_exceptions=False)
         with patch(
-            "immich_memories.ui.app.get_config",
+            "immich_memories.ui.health_api.get_config",
             side_effect=RuntimeError("configuration unavailable"),
         ):
             live = client.get("/health/live")
@@ -117,7 +117,8 @@ class TestHealthEndpoint:
 
     def test_readiness_reports_the_in_process_automation_timer(self, tmp_path: Path):
         """Docker users check /health to see when the in-app daily run will fire (#305)."""
-        from immich_memories.ui.app import _ImmichDependency, app
+        from immich_memories.ui.app import app
+        from immich_memories.ui.health_api import _ImmichDependency
 
         config = Config(
             immich={"url": "http://immich.test", "api_key": "health-secret"},
@@ -134,13 +135,13 @@ class TestHealthEndpoint:
         client = TestClient(app, raise_server_exceptions=False)
         with (
             # WHY: /health must not reach out to a real Immich in a unit test.
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(status="ready", reachable=True),
             ),
-            patch("immich_memories.ui.app.automation_scheduler", scheduler),
+            patch("immich_memories.ui.health_api.automation_scheduler", scheduler),
         ):
             response = client.get("/health/ready")
 
@@ -152,14 +153,17 @@ class TestHealthEndpoint:
 
     def test_registered_readiness_uses_one_configuration_snapshot(self):
         """A real detailed request cannot reload config while reading run history."""
-        from immich_memories.ui.app import _ImmichDependency, app
+        from immich_memories.ui.app import app
+        from immich_memories.ui.health_api import _ImmichDependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         client = TestClient(app, raise_server_exceptions=False)
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config) as get_config_mock,
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api.get_config", return_value=config
+            ) as get_config_mock,
+            patch(
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(
                     status="ready",
@@ -167,7 +171,7 @@ class TestHealthEndpoint:
                     resolved_api_version="v3",
                 ),
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
             patch("immich_memories.tracking.run_database.RunDatabase") as database,
         ):
             database.return_value.list_runs.return_value = []
@@ -194,16 +198,16 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_readiness_projects_real_automation_delivery_defaults(self, tmp_path: Path):
         """The detailed endpoint consumes delivery fields from the real status producer."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = Config(
             immich={"url": "http://immich.test", "api_key": "health-secret"},
             cache={"database": str(tmp_path / "health.db"), "directory": str(tmp_path / "cache")},
         )
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(
                     status="ready",
@@ -211,7 +215,7 @@ class TestHealthEndpoint:
                     resolved_api_version="v3",
                 ),
             ),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -226,7 +230,7 @@ class TestHealthEndpoint:
             NotificationFailureCategory,
             NotificationStateStore,
         )
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = Config(
             immich={"url": "http://immich.test", "api_key": "health-secret"},
@@ -237,15 +241,15 @@ class TestHealthEndpoint:
             NotificationFailureCategory.QUOTA
         )
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(
                     status="ready", reachable=True, resolved_api_version="v3"
                 ),
             ),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -257,17 +261,17 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_liveness_is_process_only(self):
         """Liveness must stay green without consulting config, Immich, or SQLite."""
-        from immich_memories.ui.app import _liveness_handler
+        from immich_memories.ui.health_api import _liveness_handler
 
         config = MagicMock(name="get_config")
         immich = MagicMock(name="check_immich_dependency")
         automation = MagicMock(name="get_automation_status")
         last_run = MagicMock(name="get_last_successful_run")
         with (
-            patch("immich_memories.ui.app.get_config", config),
-            patch("immich_memories.ui.app._check_immich_dependency", immich),
-            patch("immich_memories.ui.app._get_automation_status", automation),
-            patch("immich_memories.ui.app._get_last_successful_run", last_run),
+            patch("immich_memories.ui.health_api.get_config", config),
+            patch("immich_memories.ui.health_api._check_immich_dependency", immich),
+            patch("immich_memories.ui.health_api._get_automation_status", automation),
+            patch("immich_memories.ui.health_api._get_last_successful_run", last_run),
         ):
             response = await _liveness_handler(MagicMock())
 
@@ -292,7 +296,7 @@ class TestHealthEndpoint:
         failing_boundary,
     ):
         """Database diagnostics must be sanitized before response or logging boundaries."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = _config_with_every_secret_class()
         secrets = configured_secret_values(config)
@@ -304,11 +308,11 @@ class TestHealthEndpoint:
         else:
             last_run.side_effect = failure
 
-        caplog.set_level(logging.WARNING, logger="immich_memories.ui.app")
+        caplog.set_level(logging.WARNING, logger="immich_memories.ui.health_api")
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(
                     status="ready",
@@ -316,8 +320,8 @@ class TestHealthEndpoint:
                     resolved_api_version="v3",
                 ),
             ),
-            patch("immich_memories.ui.app._get_automation_status", automation),
-            patch("immich_memories.ui.app._get_last_successful_run", last_run),
+            patch("immich_memories.ui.health_api._get_automation_status", automation),
+            patch("immich_memories.ui.health_api._get_last_successful_run", last_run),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -330,16 +334,16 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_readiness_is_503_when_configuration_is_missing(self):
         """A default config cannot generate memories and must not be reported ready."""
-        from immich_memories.ui.app import _readiness_handler
+        from immich_memories.ui.health_api import _readiness_handler
 
         with (
-            patch("immich_memories.ui.app.get_config", return_value=Config()),
+            patch("immich_memories.ui.health_api.get_config", return_value=Config()),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 side_effect=AssertionError("missing config should not contact Immich"),
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -353,19 +357,19 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_readiness_is_503_when_immich_is_unreachable(self):
         """An unreachable configured server must fail the readiness gate."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         dependency = _ImmichDependency(status="unreachable", reachable=False)
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=dependency,
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -375,19 +379,19 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_readiness_is_503_for_unsupported_immich(self):
         """Auto-detecting an unsupported major must fail the readiness gate."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         dependency = _ImmichDependency(status="unsupported_version", reachable=True)
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=dependency,
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -397,7 +401,7 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_ready_snapshot_exposes_policy_version_and_automation_without_secrets(self):
         """Operators get actionable state, while configured credentials never enter JSON."""
-        from immich_memories.ui.app import _ImmichDependency, _readiness_handler
+        from immich_memories.ui.health_api import _ImmichDependency, _readiness_handler
 
         secret = "never-serialize-this-key"  # noqa: S105
         config = Config(immich={"url": "http://immich.test", "api_key": secret})
@@ -418,15 +422,15 @@ class TestHealthEndpoint:
             },
         }
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=dependency,
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=automation),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=automation),
             patch(
-                "immich_memories.ui.app._get_last_successful_run",
+                "immich_memories.ui.health_api._get_last_successful_run",
                 return_value="2026-08-11T06:15:00+00:00",
             ),
         ):
@@ -457,19 +461,19 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_legacy_health_keeps_detailed_payload_and_http_200_when_degraded(self):
         """Compatibility clients retain details without inheriting readiness status codes."""
-        from immich_memories.ui.app import _health_handler, _ImmichDependency
+        from immich_memories.ui.health_api import _health_handler, _ImmichDependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         dependency = _ImmichDependency(status="unreachable", reachable=False)
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=dependency,
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _health_handler(MagicMock())
 
@@ -481,19 +485,19 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_legacy_health_keeps_ok_status_when_dependency_is_healthy(self):
         """Compatibility health keeps its established healthy status label."""
-        from immich_memories.ui.app import _health_handler, _ImmichDependency
+        from immich_memories.ui.health_api import _health_handler, _ImmichDependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         dependency = _ImmichDependency(status="ready", reachable=True, resolved_api_version="v3")
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=dependency,
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _health_handler(MagicMock())
 
@@ -503,15 +507,15 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_health_handler_returns_json(self):
         """Health handler should return valid JSON with expected keys."""
-        from immich_memories.ui.app import _health_handler
+        from immich_memories.ui.health_api import _health_handler
 
         mock_request = MagicMock()
 
         # WHY: detailed health reads both disk configuration and SQLite status.
         with (
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
-            patch("immich_memories.ui.app.get_config", return_value=Config()),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api.get_config", return_value=Config()),
         ):
             response = await _health_handler(mock_request)
 
@@ -524,7 +528,7 @@ class TestHealthEndpoint:
     async def test_dependency_probe_resolves_version_and_authenticates(self):
         """A usable connection reports the selected API compatibility version."""
         from immich_memories.api.compatibility import ResolvedApiVersion
-        from immich_memories.ui.app import _check_immich_dependency
+        from immich_memories.ui.health_api import _check_immich_dependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         client = AsyncMock()
@@ -543,7 +547,7 @@ class TestHealthEndpoint:
     async def test_dependency_probe_accepts_a_supported_v2_server(self):
         """A configured v2 server is just as ready as the existing v3 path."""
         from immich_memories.api.compatibility import ResolvedApiVersion
-        from immich_memories.ui.app import _check_immich_dependency
+        from immich_memories.ui.health_api import _check_immich_dependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         client = AsyncMock()
@@ -561,7 +565,7 @@ class TestHealthEndpoint:
         """A rejected key means the server answered, but it is not ready for work."""
         from immich_memories.api.compatibility import ResolvedApiVersion
         from immich_memories.api.immich import ImmichAuthError
-        from immich_memories.ui.app import _check_immich_dependency
+        from immich_memories.ui.health_api import _check_immich_dependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         client = AsyncMock()
@@ -578,7 +582,7 @@ class TestHealthEndpoint:
     @pytest.mark.asyncio
     async def test_readiness_timeout_returns_503_and_closes_cancelled_client(self):
         """A cancelled dependency probe releases its client and degrades readiness."""
-        from immich_memories.ui.app import _readiness_handler
+        from immich_memories.ui.health_api import _readiness_handler
 
         async def cancelled_wait_for(operation, timeout):  # noqa: ARG001
             task = asyncio.create_task(operation)
@@ -591,11 +595,11 @@ class TestHealthEndpoint:
         client = AsyncMock()
         client.__aenter__.return_value = client
         with (
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch("immich_memories.api.immich.ImmichClient", return_value=client),
-            patch("immich_memories.ui.app.asyncio.wait_for", new=cancelled_wait_for),
-            patch("immich_memories.ui.app._get_automation_status", return_value=None),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value=None),
+            patch("immich_memories.ui.health_api.asyncio.wait_for", new=cancelled_wait_for),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=None),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value=None),
         ):
             response = await _readiness_handler(MagicMock())
 
@@ -607,7 +611,7 @@ class TestHealthEndpoint:
     async def test_dependency_probe_distinguishes_an_unsupported_server(self):
         """A responding unsupported major is different from a network outage."""
         from immich_memories.api.compatibility import UnsupportedImmichVersion
-        from immich_memories.ui.app import _check_immich_dependency
+        from immich_memories.ui.health_api import _check_immich_dependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         client = AsyncMock()
@@ -621,7 +625,7 @@ class TestHealthEndpoint:
 
     def test_get_last_successful_run_returns_none_when_no_runs(self):
         """Should return None when no completed runs exist."""
-        from immich_memories.ui.app import _get_last_successful_run
+        from immich_memories.ui.health_api import _get_last_successful_run
 
         # WHY: RunDatabase reads from SQLite file (lazy import inside function)
         with patch("immich_memories.tracking.run_database.RunDatabase") as mock_db_cls:
@@ -636,7 +640,8 @@ class TestHealthDisclosure:
 
     @pytest.mark.parametrize("path", ["/health", "/health/ready"])
     def test_unauthenticated_health_omits_automation_detail_when_auth_enabled(self, path: str):
-        from immich_memories.ui.app import _ImmichDependency, app
+        from immich_memories.ui.app import app
+        from immich_memories.ui.health_api import _ImmichDependency
 
         config = Config(
             immich={"url": "http://immich.test", "api_key": "health-secret"},
@@ -655,14 +660,14 @@ class TestHealthDisclosure:
         client = TestClient(app, raise_server_exceptions=False)
         with (
             # WHY: config and Immich probe are external boundaries; the test is about payload shape.
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(status="ready", reachable=True),
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=automation),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value="run-123"),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=automation),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value="run-123"),
         ):
             response = client.get(path)
 
@@ -676,21 +681,22 @@ class TestHealthDisclosure:
         assert body["last_successful_run"] is None
 
     def test_health_keeps_detail_when_auth_disabled(self):
-        from immich_memories.ui.app import _ImmichDependency, app
+        from immich_memories.ui.app import app
+        from immich_memories.ui.health_api import _ImmichDependency
 
         config = Config(immich={"url": "http://immich.test", "api_key": "health-secret"})
         automation = {"last_attempt": {"memory_key": "year_in_review:2024"}}
         client = TestClient(app, raise_server_exceptions=False)
         with (
             # WHY: same boundaries as above; auth disabled means a trusted LAN deployment.
-            patch("immich_memories.ui.app.get_config", return_value=config),
+            patch("immich_memories.ui.health_api.get_config", return_value=config),
             patch(
-                "immich_memories.ui.app._check_immich_dependency",
+                "immich_memories.ui.health_api._check_immich_dependency",
                 new_callable=AsyncMock,
                 return_value=_ImmichDependency(status="ready", reachable=True),
             ),
-            patch("immich_memories.ui.app._get_automation_status", return_value=automation),
-            patch("immich_memories.ui.app._get_last_successful_run", return_value="run-123"),
+            patch("immich_memories.ui.health_api._get_automation_status", return_value=automation),
+            patch("immich_memories.ui.health_api._get_last_successful_run", return_value="run-123"),
         ):
             response = client.get("/health")
 
@@ -710,9 +716,9 @@ class TestHealthIsCheapUnderRepeatedProbes:
 
     @staticmethod
     def _reset_cache() -> None:
-        from immich_memories.ui import app as app_module
+        from immich_memories.ui import health_api as health_module
 
-        app_module._health_snapshot_cache = None
+        health_module._health_snapshot_cache = None
 
     @staticmethod
     def _configured_config():
@@ -728,55 +734,55 @@ class TestHealthIsCheapUnderRepeatedProbes:
 
     @pytest.mark.asyncio
     async def test_a_burst_of_probes_does_the_work_once(self):
-        from immich_memories.ui import app as app_module
+        from immich_memories.ui import health_api as health_module
 
         self._reset_cache()
         calls = []
 
         async def counted_dependency(config):  # noqa: ARG001
             calls.append(1)
-            return app_module._ImmichDependency(status="ready", reachable=True)
+            return health_module._ImmichDependency(status="ready", reachable=True)
 
         # WHY: the Immich server and the SQLite stores behind the snapshot
         with (
             # WHY: the probe only runs for a configured Immich
-            patch("immich_memories.ui.app.get_config", self._configured_config),
+            patch("immich_memories.ui.health_api.get_config", self._configured_config),
             # WHY: external Immich server
-            patch("immich_memories.ui.app._check_immich_dependency", counted_dependency),
+            patch("immich_memories.ui.health_api._check_immich_dependency", counted_dependency),
             # WHY: reads four SQLite databases
-            patch("immich_memories.ui.app._operational_detail", return_value=(None, None)),
+            patch("immich_memories.ui.health_api._operational_detail", return_value=(None, None)),
         ):
             for _ in range(5):
-                await app_module._health_handler(MagicMock())
+                await health_module._health_handler(MagicMock())
 
         assert len(calls) == 1, f"{len(calls)} dependency probes for 5 requests"
 
     @pytest.mark.asyncio
     async def test_the_snapshot_goes_stale_so_readiness_stays_truthful(self, monkeypatch):
         """A cache that never expires would report a dead Immich as ready."""
-        from immich_memories.ui import app as app_module
+        from immich_memories.ui import health_api as health_module
 
         self._reset_cache()
         calls = []
         clock = {"now": 1000.0}
-        monkeypatch.setattr(app_module.time, "monotonic", lambda: clock["now"])
+        monkeypatch.setattr(health_module.time, "monotonic", lambda: clock["now"])
 
         async def counted_dependency(config):  # noqa: ARG001
             calls.append(1)
-            return app_module._ImmichDependency(status="ready", reachable=True)
+            return health_module._ImmichDependency(status="ready", reachable=True)
 
         # WHY: the Immich server and the SQLite stores behind the snapshot
         with (
             # WHY: the probe only runs for a configured Immich
-            patch("immich_memories.ui.app.get_config", self._configured_config),
+            patch("immich_memories.ui.health_api.get_config", self._configured_config),
             # WHY: external Immich server
-            patch("immich_memories.ui.app._check_immich_dependency", counted_dependency),
+            patch("immich_memories.ui.health_api._check_immich_dependency", counted_dependency),
             # WHY: reads four SQLite databases
-            patch("immich_memories.ui.app._operational_detail", return_value=(None, None)),
+            patch("immich_memories.ui.health_api._operational_detail", return_value=(None, None)),
         ):
-            await app_module._health_handler(MagicMock())
+            await health_module._health_handler(MagicMock())
             clock["now"] += 60.0
-            await app_module._health_handler(MagicMock())
+            await health_module._health_handler(MagicMock())
 
         assert len(calls) == 2, "the snapshot never refreshed"
 
@@ -786,7 +792,7 @@ class TestHealthIsCheapUnderRepeatedProbes:
         import asyncio
         import time as time_module
 
-        from immich_memories.ui import app as app_module
+        from immich_memories.ui import health_api as health_module
 
         self._reset_cache()
         ticks = 0
@@ -804,21 +810,21 @@ class TestHealthIsCheapUnderRepeatedProbes:
             return None, None
 
         async def ready_dependency(config):  # noqa: ARG001
-            return app_module._ImmichDependency(status="ready", reachable=True)
+            return health_module._ImmichDependency(status="ready", reachable=True)
 
         # WHY: stands in for the Immich server and for a slow SQLite read
         with (
             # WHY: the probe only runs for a configured Immich
-            patch("immich_memories.ui.app.get_config", self._configured_config),
+            patch("immich_memories.ui.health_api.get_config", self._configured_config),
             # WHY: external Immich server
-            patch("immich_memories.ui.app._check_immich_dependency", ready_dependency),
+            patch("immich_memories.ui.health_api._check_immich_dependency", ready_dependency),
             # WHY: a real contended SQLite read, without needing contention
-            patch("immich_memories.ui.app._operational_detail", slow_blocking_detail),
+            patch("immich_memories.ui.health_api._operational_detail", slow_blocking_detail),
         ):
             ticker = asyncio.create_task(tick())
             await asyncio.sleep(0.01)
             before = ticks
-            await app_module._health_handler(MagicMock())
+            await health_module._health_handler(MagicMock())
             # Sampled here, not after awaiting the ticker: once the handler
             # returns the ticker finishes either way, which would make this
             # pass whether or not the loop was ever blocked.
